scripts/requirements.txt
docling>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
pymupdf>=1.24.0
